UPDATE LISTLAST FOR CIRCULAR LISTS:
Updating ft_listlast to include check on circular lists.
The idea of a ft_listlast function for a circular list may seem weird.
Therefore the result obtained will be the last list element before
the next one is actually the same as start.

Another way of doing it would have been to check the previous element.
Except this only works for lists including a prev address.